+++ backend/test/support/mocks.ex
@@ -1,7 +1,7 @@
 #
 # This file is part of Edgehog.
 #
-# Copyright 2021-2023 SECO Mind Srl
+# Copyright 2021 - 2025 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -80,6 +80,10 @@ Mox.defmock(Edgehog.Astarte.Device.ForwarderSessionMock,
 
 Mox.defmock(Edgehog.Astarte.Interface.MockDataLayer,
   for: Edgehog.Astarte.Interface.DataLayer
+)
+
+Mox.defmock(Edgehog.Astarte.DeliveryPolicies.MockDataLayer,
+  for: Edgehog.Astarte.DeliveryPolicies.DataLayer
 )
 
 Mox.defmock(Edgehog.Astarte.Trigger.MockDataLayer,